+++ src/libmerc/dnp3.h
@@ -525,6 +525,12 @@ public:
         return;
     }
 
+    void write_l7_metadata(cbor_object &o, bool) {
+        cbor_array protocols{o, "protocols"};
+        protocols.print_string("dnp3");
+        protocols.close();
+    }
+
 };
 
 namespace {